requirements.txt
channels==4.1.1
channels-redis==4.2.0
daphne==4.1.2
Django==4.2.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
python-dotenv==1.0.1
redis==5.0.4

